Set up webpack for sass compilation and js bundling.
This commit enables us to use webpack to bundle our javascript files
and to compile our .scss files to .css files. This is necessary since
we will be writing sass going forward. We also introduce a new
"watch" command for developing, which can be run via: "npm run watch"
in a separate tab. "Watch" will start a process that watches updates
to our .scss and .js files and will keep updating the bundle present
in the build folder. Henceforth, all references to js files should use
the path `/bundle/*` when referenced from .html files (we only want to
serve the built bundles to the user). This commit also makes some
minor code changes to the server because of the change in the paths of
our css and js files.